requirements.txt
Django<4.3,>=4.2
argon2-cffi>=19.1.0
django-environ==0.10.0
djangorestframework<3.15,>=3.14.0
# PostgreSQL driver
psycopg<3.2,>=3.1.8
# Requires Python development files and libpq-dev to build
psycopg-c<3.2,>=3.1.6
psycopg-pool
django-filter<23.2,>=23.1
drf-spectacular==0.26.1
dj-rest-auth<3.1,>=3.0.0
django-allauth<0.53.0,>=0.40.0
